Implement locking to allow multiple writers
To keep lock contention to a minimum, writes are done to independent temporary
files. Exclusive locks are only held while replacing the "final" files. Shared
locks are held whenever the files are opened (only for the duration of the open
operation itself) to ensure consistency between the log file and hash file.