Rescue all exceptions when trying to clean up an empty directory
In rare cases we were seeing ENOENT and ENOTDIR in addition to ENOTEMPTY.
ENOENT probably happens if another process removes the directory first -- this
is possible since we don't take a lock between the close and remove operations.
ENOTDIR happens when the hash gets opened with "current" instead of the
resolved path. It is unclear why this happens, but the problem is at open-time
and we probably shouldn't throw an exception at close-time if it happens.