Implement more methods in HashMethods; move flatten to HashMethods
Now a backend need only implement four methods: [], []=, delete, and each (and
close, but that's not a hash method). However, a backend may choose to override
some of the default implementations if the backend is able to implement the
methods more efficiently. For example, the Sparkey backend overrides clear,
has_key?, keys, replace, size, and values.