Move HashMethods to the top level
Although it sort of defines an interface for backends to implement, it didn't
really belong under backend. It also doesn't really belong in the frontend (or
the Hammerspace::Hash namespace) because the frontend delegates to a backend
that mixes these in.

I see it like Enumerable, more like a set of utility methods than anything
else.