Fix a ruby crash on OS X
There's a bug in gnista 0.0.4 where if you try to open a hash reader and that
raises an exception, the next GC call will cause ruby to crash. This was
observed on OS X with ruby 1.9.3; here's a minimal test case:

    require 'gnista'

    begin
      Gnista::Hash.new(rand.to_s, rand.to_s)
    rescue GnistaException
    end

    GC.start

Adding checks to ensure the index and log files exist before attempting to open
them works around the most common cause of this issue.