Avoid an unnecessary call to Gnista::Hash#include? on get
Previously we called include? to decide whether or not to return the default
value instead of just checking the value. This ensures that a stored false or
nil value would not fall back to the default. However, this is unnecessary
because false and nil cannot be stored in Gnista hashes, only strings are
allowed and strings are truthy in ruby.